backend/requirements.txt
flask
flask-cors
flask-socketio
tweepy
nltk
gensim
transformers
torch
requests
python-dotenv
